split structs into different files and apply review comments
